requirements.txt
Flask>=2.0.1
Flask-Cors>=3.0
Flask-SQLAlchemy>=3.1.1
Flask-RESTful>=0.3.10
flask-marshmallow>=0.14.0
marshmallow-sqlalchemy>=1.0.0
SQLAlchemy>=2.0.25
flask-restplus>=0.9.2
MarkupSafe>=2.0.1
Jinja2>=3.0.1
python-dotenv>=1.0.0
Werkzeug>=2.0.1
flasgger>=0.9.5
Flask-JWT-Extended>=4.2.3
PyJWT>=2.1.0
PyYAML>=5.4.1